Use a List of Custom Programming Event Objects to hold code.
The exercise wants you to use a HashMap for something a HashMap
isn't supposed to be used for (a list where all values are
meaningful and need to be accessed like a list. So create a
ProgrammingEvent object and store data in that. This this
also considerably easier at extension than a HashMap is.